+++ src/app/[orgId]/[locationId]/manage/_components/AddMenuDialog.tsx
@@ -9,10 +9,9 @@ import {
   DialogTitle,
 } from "~/components/ui/dialog";
 import { AddMenu } from "./AddMenu";
-import { type OrgTier } from "~/app/_domain/price-tiers";
 import { NoQuotaLeft } from "../../_components/NoQuotaLeft";
 
-export default function AddMenuDialog(props: { orgTier: OrgTier }) {
+export default function AddMenuDialog(props: { availableQuota: number }) {
   const router = useRouter();
 
   const handleOpenChange = (open: boolean) => {
@@ -20,33 +19,30 @@ export default function AddMenuDialog(props: { orgTier: OrgTier }) {
       router.back();
     }
   };
+  if (props.availableQuota <= 0) {
+    return (
+      <Dialog open={true} onOpenChange={handleOpenChange}>
+        <DialogContent>
+          <DialogHeader>
+            <DialogTitle>Add a menu</DialogTitle>
+            <DialogDescription> </DialogDescription>
+          </DialogHeader>
+          <NoQuotaLeft title={"Sorry, you ran out of menus..."} />
+        </DialogContent>
+      </Dialog>
+    );
+  }
 
   return (
     <Dialog open={true} onOpenChange={handleOpenChange}>
       <DialogContent>
-        {props.orgTier.available > 0 && (
-          <>
-            <DialogHeader>
-              <DialogTitle>Add a menu</DialogTitle>
-              <DialogDescription>
-                Fill the form below to create a menu.
-              </DialogDescription>
-            </DialogHeader>
-            <AddMenu  />
-          </>
-        )}
-        {props.orgTier.available <= 0 && (
-          <>
-            <DialogHeader>
-              <DialogTitle>Add a menu</DialogTitle>
-              <DialogDescription>
-                 {' '}
-              </DialogDescription>
-            </DialogHeader>
-            
-            <NoQuotaLeft title={"You ran out of menus..."} orgTier={props.orgTier} />
-          </>
-        )}
+        <DialogHeader>
+          <DialogTitle>Add a menu</DialogTitle>
+          <DialogDescription>
+            Fill the form below to create a menu.
+          </DialogDescription>
+        </DialogHeader>
+        <AddMenu />
       </DialogContent>
     </Dialog>
   );