+++ src/app/[orgId]/[locationId]/manage/menus/add/page.tsx
@@ -1,14 +1,6 @@
-import { auth } from "@clerk/nextjs/server";
 import AddMenuDialog from "../../_components/AddMenuDialog";
-import {
-  defaultTier,
-  type OrgTier,
-  PriceTierIdSchema,
-  priceTiers,
-} from "~/app/_domain/price-tiers";
-import { getMenusByLocation } from "~/server/queries";
 import { locationIdSchema } from "~/app/_domain/location";
-import { BoxError } from "~/app/_components/BoxError";
+import { getAvailableQuota } from "~/app/_utils/quota-utils";
 
 type Params = Promise<{ locationId: string }>;
 
@@ -16,28 +8,11 @@ export default async function AddMenuPage(props: { params: Params }) {
   const params = await props.params;
   const validationResult = locationIdSchema.safeParse(params.locationId);
   if (!validationResult.success) {
-    return <BoxError errorTypeId={"MENUS_INVALID_PARAM"} />;
+    // TODO Test
+    throw new Error("Location issue");
   }
-  const parsedLocationId = validationResult.data;
+ 
+  const availableQuota = await getAvailableQuota('menus');
 
-  const priceTierId = (await auth()).sessionClaims?.metadata?.tier;
-  const parsedTier = PriceTierIdSchema.safeParse(priceTierId);
-  // TODO revisit. or maybe just always trust JWT token
-  const parsedOrDefaultTier = parsedTier.success
-    ? parsedTier.data
-    : defaultTier;
-
-  const menus = await getMenusByLocation(parsedLocationId);
-
-  const quota = priceTiers[parsedOrDefaultTier].menus;
-  const orgTier: OrgTier = {
-    priceTierId: parsedOrDefaultTier,
-    resourceSingularName: 'menu',
-    resourcePluralName: 'menus',
-    quota,
-    used: menus.length,
-    available: quota - menus.length,
-  };
-
-  return <AddMenuDialog orgTier={orgTier} />;
+  return <AddMenuDialog availableQuota={availableQuota} />;
 }