+++ src/app/[orgId]/[locationId]/reports/page.tsx
@@ -1,9 +1,6 @@
 import Link from "next/link";
-import { BoxError } from "~/app/_components/BoxError";
 import { locationIdSchema } from "~/app/_domain/location";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "~/components/ui/tabs";
-
- 
 
 type Params = Promise<{ locationId: string }>;
 
@@ -12,7 +9,9 @@ type Params = Promise<{ locationId: string }>;
 
   const validationResult = locationIdSchema.safeParse(params.locationId);
   if (!validationResult.success) {
-    return <BoxError errorTypeId={"MENUS_INVALID_PARAM"} />;
+    // TODO new error component
+    return 
+    //return <BoxError errorTypeId={"MENUS_INVALID_PARAM"} />;
   }
 
   //TODO replace with actual functionality