+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
+## 0.0.33
+
+### Features
+- View plan details (/view-plan)
+- Price tier quota calculations for locations and menus
+
 ## 0.0.32
 
 ### Features
-- Change Plan
+- Users can now change their price tier (/change-plan)